% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_bbbq_1.R
\name{answer_bbbq_1}
\alias{answer_bbbq_1}
\title{Answer the BBBQ}
\usage{
answer_bbbq_1(
  fasta_filename,
  alleles = "DRB1_0101",
  binding_strength_threshold = 5,
  pureseqtmr_filename = tempfile(".txt"),
  netmhc2pan_filename = tempfile(".csv"),
  epitopeome_filename = tempfile(".fasta")
)
}
\arguments{
\item{fasta_filename}{name of a FASTA file}

\item{alleles}{one or more alleles, e.g. \code{DRB1_0101}.
See \link[netmhc2pan]{get_netmhc2pan_alleles} for a full list}

\item{binding_strength_threshold}{threshold of the MHC2 allele's
binding strength. A low threshold, e.g. \code{5.0}, selects for the top
5 percent best binders}

\item{pureseqtmr_filename}{name of the file to store the
TMHMM output.
By default, this filename is generated by \link{tempfile},
and will be put in the \code{/tmp} folder,
which is cleaned by the operating system}

\item{netmhc2pan_filename}{name of the file to store the
NetMHC2pan output.
By default, this filename is generated by \link{tempfile},
and will be put in the \code{/tmp} folder,
which is cleaned by the operating system}

\item{epitopeome_filename}{name of the file to store the epitopeome
output.
By default, this filename is generated by \link{tempfile},
and will be put in the \code{/tmp} folder,
which is cleaned by the operating system}
}
\value{
a data frame with counts of the epitopa
}
\description{
Answer the BBBQ
}
\author{
Richèl J.C. Bilderbeek
}
