% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_detected_by_mhc_1.R
\name{are_detected_by_mhc_1}
\alias{are_detected_by_mhc_1}
\title{Determine if the protein sequences are detected by MHC-I}
\usage{
are_detected_by_mhc_1(protein_sequences)
}
\arguments{
\item{protein_sequences}{one or more protein sequences}
}
\description{
Determine if the protein sequences are detected by MHC-I
}
