% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_detected_by_mhc_2.R
\name{are_detected_by_mhc_2}
\alias{are_detected_by_mhc_2}
\title{Determine if each protein sequence has one MHC-II binder}
\usage{
are_detected_by_mhc_2(protein_sequences)
}
\arguments{
\item{protein_sequences}{one or more protein sequences}
}
\description{
Determine if each protein sequence has one MHC-II binder
}
