% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_overlap.R
\name{calc_overlap}
\alias{calc_overlap}
\title{Calculate the fractions of MHC epitopes that overlap with a TMH
for multiple MHC haplotypes for one type of agent.}
\usage{
calc_overlap(target_name, haplotype, n_aas)
}
\arguments{
\item{n_aas}{number of amino acids}
}
\value{
per haplotype, the fraction of MHC epitopes
that overlap with a TMH. The fractions range from 0 ('none')
to (and including) 1 ('all')
}
\description{
Calculate the fractions of MHC epitopes that overlap with a TMH
for multiple MHC haplotypes for one type of agent.
}
\author{
Richèl J.C. Bilderbeek
}
