% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_overlap_peptide.R
\name{calc_overlap_peptide}
\alias{calc_overlap_peptide}
\title{Predict the number of binders and the number of binders
that overlap with at least one amino acid with a TMH.}
\usage{
calc_overlap_peptide(
  peptide,
  target_name,
  haplotype,
  n_aas,
  percentile = 0.02,
  verbose = FALSE
)
}
\arguments{
\item{n_aas}{number of amino acids}

\item{percentile}{how low the IC50 must be for the protein to
be considered a binder. For example, 0.02 denotes that the protein
must have an IC50 in the lowest 2 percent range}

\item{verbose}{set to TRUE for more output}
}
\value{
a tibble with
\enumerate{
  \item n_binders number of epitopes that bind
  \item n_binders_tmh number of epitopes that bind and have one
    amino acid overlapping with a TMH
}
}
\description{
Predict the number of binders and the number of binders
that overlap with at least one amino acid with a TMH.
}
\examples{
peptide <- "MYSFVSEETGTLIVNSVLLFLAFVVFLLVTLAILTALRLCAYCCNIVNVSLVKPSFYVYSRVKN"

calc_overlap_peptide(
  peptide = peptide,
  target_name = "test",
  haplotype = get_mhc1_haplotypes()[1],
  n_aas = 9,
)
}
\author{
Richèl J.C. Bilderbeek
}
