% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_overlaps.R
\name{calc_overlaps}
\alias{calc_overlaps}
\title{Calculate the fractions of MHC epitopes that overlap with a TMH.}
\usage{
calc_overlaps(proteome_filename, haplotypes)
}
\arguments{
\item{proteome_filename}{name of the file the proteome will be stored at}

\item{haplotypes}{a list of MHC-I or MHC-II haplotypes}
}
\value{
per haplotype, the fraction of MHC epitopes
that overlap with a TMH. The fractions range from 0 ('none')
to (and including) 1 ('all')
}
\description{
Per haplotype, calculate the fraction of MHC epitopes
that overlap with a TMH.
}
\author{
Richèl J.C. Bilderbeek
}
