% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_p_det_tmh_mut.R
\name{calc_p_det_tmh_mut}
\alias{calc_p_det_tmh_mut}
\title{Calculate the chance that a TMH one-mutation
mutant is detected}
\usage{
calc_p_det_tmh_mut(protein_sequence)
}
\arguments{
\item{protein_sequence}{a protein sequence}
}
\description{
Calculate the chance that a TMH one-mutation
mutant is detected
}
