% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_haplotype.R
\name{check_haplotype}
\alias{check_haplotype}
\title{Check the haplotype}
\usage{
check_haplotype(haplotype, ic50_prediction_tool)
}
\arguments{
\item{haplotype}{one MHC-I or MHC-II haplotype}

\item{ic50_prediction_tool}{tool to predict the IC50 from a
  peptide. Possible values are:\cr
\itemize{
  \item mhcnuggetsr \link[mhcnuggetsr]{mhcnuggetsr},
    which uses MHCnuggets
  \item mhcnuggetsr \link[netmhc2pan]{netmhc2pan},
    which uses NetMHC2pam
  \item EpitopePrediction uses \code{EpitopePrediction}
}}
}
\description{
Check if the haplotype can be converted to a valid
}
\details{
Will \link{stop} if the name is invalid
}
\author{
Richèl J.C. Bilderbeek
}
