% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ic50_prediction_tool.R
\name{check_ic50_prediction_tool}
\alias{check_ic50_prediction_tool}
\title{Check the name of the IC50 prediction tool}
\usage{
check_ic50_prediction_tool(ic50_prediction_tool)
}
\description{
Check the name of the IC50 prediction tool.
Will \link{stop} if the name is invalid
}
\author{
Richèl J.C. Bilderbeek
}
