% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mhc_1_haplotype_name.R
\name{check_mhc_1_haplotype_name}
\alias{check_mhc_1_haplotype_name}
\title{Check if the name is valid}
\usage{
check_mhc_1_haplotype_name(mhc_1_haplotype)
}
\arguments{
\item{mhc_1_haplotype}{HMC-I haplotope.
Use \link{get_mhc1_haplotypes} to get a list of all haplotypes}
}
\description{
Check if the name is valid
}
