% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mhc_haplotype_name.R
\name{check_mhc_haplotype_name}
\alias{check_mhc_haplotype_name}
\title{Check an MHC haplotype name}
\usage{
check_mhc_haplotype_name(mhc_haplotype)
}
\arguments{
\item{mhc_haplotype}{haplotype of either HMC-I or MHC-II.
Use \link{get_mhc1_haplotypes} to get a list of all MHC-I haplotypes.
Use \link{get_mhc2_haplotypes} to get a list of all MHC-II haplotypes.}
}
\description{
Check an MHC haplotype name
}
