% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_protein_sequences_length.R
\name{check_protein_sequences_length}
\alias{check_protein_sequences_length}
\title{If if the protein sequences are long enough}
\usage{
check_protein_sequences_length(protein_sequences, peptide_length)
}
\arguments{
\item{protein_sequences}{one or more protein sequences}

\item{peptide_length}{length of the peptide in amino acids}
}
\value{
Nothing
}
\description{
If if the protein sequences are longer than the
desired epitope peptide length: to estimate epitope binding
strength using, for example, 13-mers, the protein sequences
must have a least 13 amino acids
}
\author{
Richèl J.C. Bilderbeek
}
