% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_proteome_type.R
\name{check_proteome_type}
\alias{check_proteome_type}
\title{Check if a \code{proteome_type} is valid}
\usage{
check_proteome_type(proteome_type)
}
\arguments{
\item{proteome_type}{type of proteome, must be
either \code{full} or \code{representative}}
}
\description{
Check if a \code{proteome_type} is valid.
Will \link{stop} if not.
}
