% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_target_name.R
\name{check_target_name}
\alias{check_target_name}
\title{Check if the target name is valid}
\usage{
check_target_name(target_name)
}
\arguments{
\item{target_name}{abbreviated name of organism or virus.
Use \link{get_target_names} to get the valid values.
Use \link{get_target_english_name} to get the English name of a target.}
}
\description{
Check if the target name is valid.
Will \link{stop} if not
}
\seealso{
Use \link{get_target_english_name} to get the English name of a target
}
