% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_topology_prediction_tool.R
\name{check_topology_prediction_tool}
\alias{check_topology_prediction_tool}
\title{Check the name of the topology prediction tool}
\usage{
check_topology_prediction_tool(topology_prediction_tool)
}
\description{
Check the name of the topology prediction tool.
Will \link{stop} if the name is invalid
}
\author{
Richèl J.C. Bilderbeek
}
