% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_triplet_to_amino_acid.R
\name{convert_triplet_to_amino_acid}
\alias{convert_triplet_to_amino_acid}
\title{Convert a (DNA) triplet to the corresponding amino acid}
\usage{
convert_triplet_to_amino_acid(triplet)
}
\arguments{
\item{triplet}{a DNA triplet}
}
\description{
Convert a (DNA) triplet to the corresponding amino acid
}
\note{
need to convert T to U for RNA triplets
}
\examples{
convert_triplet_to_amino_acid("GCT")
}
\author{
Richèl J.C. Bilderbeek
}
