% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate-to-hydrophobicity.R
\name{correlate_to_hydrophobicity}
\alias{correlate_to_hydrophobicity}
\title{Correlate to hydrophobicity}
\usage{
correlate_to_hydrophobicity()
}
\description{
Correlate to hydrophobicity
}
\author{
Richel J.C. Bilderbeek, adapted from Johannes Textor
}
