% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_consensus_topology_conservation.R
\name{create_consensus_topology_conservation}
\alias{create_consensus_topology_conservation}
\title{Create a consensus-topology-conservation table}
\usage{
create_consensus_topology_conservation(
  protein_sequences,
  topology_prediction_tool,
  msa_method,
  msa_subst_matrix,
  msa_filename = NULL
)
}
\arguments{
\item{protein_sequences}{one or more protein sequences}

\item{topology_prediction_tool}{tool to predict the topology
  of a membrane protein. Possible values are:\cr
\itemize{
  \item mhcnuggetsr \link[pureseqtmr]{pureseqtmr},
    which uses PureseqTM
  \item mhcnuggetsr \link[tmhmm]{tmhmm},
    which uses TMHMM
}}

\item{msa_method}{the \code{method} argument of \link[msa]{msa}.
Valid methods are \code{ClustalW}, \code{ClustalOmega} (preferred for proteins)
and \code{Muscle}}

\item{msa_subst_matrix}{multiple sequence alignment substitution matrix,
the \code{substitutionMatrix} argument of \link[msa]{msa}.
For a protein alignment, relevant values are \code{blosum},
\code{pam} and \code{gonnet}}

\item{msa_filename}{filename to save the multiple sequence
alignment to. Use \link{NULL} if there is no desire
to store it}
}
\value{
a \link{tibble} with columns:\cr
\itemize{
  \item aa the amino acid at that position in consensus sequence
  \item score the conservation score
  \item is_tmh if that amino acid at that location is part of a
    transmembrane helix
}
}
\description{
Create a consensus-topology-conservation table
}
