% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_haplotypes_lut.R
\name{create_haplotypes_lut}
\alias{create_haplotypes_lut}
\title{Create a lookup table for haplotypes, MHC class and haplotype ID}
\usage{
create_haplotypes_lut()
}
\value{
a tibble with columns:\cr
\itemize{
  \item haplotype the haplotype in official naming,
    for example \code{HLA-A*01:01}
  \item mhc_class the MHC class, for example \code{1}
  \item haplotype_id the haplotype ID, as used in file names,
    for example \code{h1}
}
}
\description{
Create a lookup table for haplotypes, MHC class and haplotype ID
}
