% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_n_mers.R
\name{create_n_mers}
\alias{create_n_mers}
\title{Create the n-mers from a character string}
\usage{
create_n_mers(string, n)
}
\arguments{
\item{string}{a character vector with one element}

\item{n}{the length of the n-mers}
}
\value{
a character vector with all n-mers
}
\description{
Create the n-mers from a character string
}
\examples{
# Creates ABCD, BCDE, CDEF
create_n_mers(string = "ABCDEF", n = 4)
}
