% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_n_mers_tibble.R
\name{create_n_mers_tibble}
\alias{create_n_mers_tibble}
\title{Create the n-mers from one or more character strings}
\usage{
create_n_mers_tibble(strings, n)
}
\arguments{
\item{strings}{a character vector with one or more elements}

\item{n}{the length of the n-mers}
}
\value{
a \link[tibble]{tibble} with two columns:
 * \code{string} the string, as a factor
 * \code{n_mers} the n-mers per each string
}
\description{
Create the n-mers from one or more character strings
}
\examples{
create_n_mers_tibble(strings = c("AB", "CDE"), n = 2)
}
\seealso{
use \link{create_n_mers} to create
the n-mers for one string (as a character vector)
}
