% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_proteins_lut.R
\name{create_proteins_lut}
\alias{create_proteins_lut}
\title{Create a protein lookup table from a proteome}
\usage{
create_proteins_lut(proteome_filename)
}
\value{
a tibble with columns:\cr
\itemize{
  \item protein_id the protein ID,
    for example \code{p1}
  \item protein the protein name,
    for example \code{sp|P0DTD8|NS7B}
  \item sequence the protein sequence,
    for example \code{MIELSLIDFYLCFL}
}
}
\description{
Create a protein lookup table from a proteome
}
