% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_extreme_peptide.R
\name{create_random_extreme_peptide}
\alias{create_random_extreme_peptide}
\title{Create a random peptide likelier to have an
extreme hydrophobicity}
\usage{
create_random_extreme_peptide(peptide_length)
}
\arguments{
\item{peptide_length}{length of the peptide in amino acids}
}
\description{
Create a random peptide likelier to have an
extreme hydrophobicity
}
\author{
Richèl J.C. Bilderbeek
}
