% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_extreme_polypeptide.R
\name{create_random_extreme_polypeptide}
\alias{create_random_extreme_polypeptide}
\title{Create a random polypeptide likelier to have an
extreme hydrophobicity}
\usage{
create_random_extreme_polypeptide(n_aas)
}
\arguments{
\item{n_aas}{number of amino acids}
}
\description{
Create a random polypeptide likelier to have an
extreme hydrophobicity
}
