% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_hydrophobic_polypeptide.R
\name{create_random_hydrophobic_polypeptide}
\alias{create_random_hydrophobic_polypeptide}
\title{Create a random hydrophobic polypeptide}
\usage{
create_random_hydrophobic_polypeptide(n_aas)
}
\arguments{
\item{n_aas}{number of amino acids}
}
\description{
Create a random hydrophobic polypeptide
}
