% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_polypeptide.R
\name{create_random_peptide}
\alias{create_random_peptide}
\title{Create a random peptide}
\usage{
create_random_peptide(n_aas)
}
\arguments{
\item{n_aas}{number of amino acids}
}
\description{
Create a random peptide
}
\author{
Richèl J.C. Bilderbeek
}
