% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_peptide_with_hydrophobicity.R
\name{create_random_peptide_with_hydrophobicity}
\alias{create_random_peptide_with_hydrophobicity}
\title{Creates a random peptide of a certain hydrophobicity}
\usage{
create_random_peptide_with_hydrophobicity(
  peptide_length,
  min_hydrophobicity,
  max_hydrophobicity,
  n_attempts = 1000
)
}
\arguments{
\item{peptide_length}{length of the peptide in amino acids}

\item{min_hydrophobicity}{the minimal hydrophobicity}

\item{max_hydrophobicity}{the maximal hydrophobicity}

\item{n_attempts}{number of attempts before the algorithm will \link{stop}}
}
\description{
Creates a random peptide of a certain hydrophobicity
}
\examples{
library(testthat)

peptide <- create_random_peptide_with_hydrophobicity(
  peptide_length = 2,
  min_hydrophobicity = 1.0,
  max_hydrophobicity = 1.5
)
expect_true(Peptides::hydrophobicity(peptide) >= 1.0)
expect_true(Peptides::hydrophobicity(peptide) <= 1.5)
}
\author{
Richèl J.C. Bilderbeek
}
