% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_tmh.R
\name{create_random_tmh}
\alias{create_random_tmh}
\title{Create a random TMH sequences}
\usage{
create_random_tmh(
  n_aas,
  create_random_polypeptide_fun = bbbq::create_random_hydrophobic_polypeptide
)
}
\arguments{
\item{n_aas}{number of amino acids}

\item{create_random_polypeptide_fun}{function to create a random
peptide fragment of the desired length}
}
\description{
Create a random TMH sequences
}
