% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tmh_fasta.R
\name{create_tmh_fasta}
\alias{create_tmh_fasta}
\title{Creates a FASTA file with the sequences of the proteome that
are TMHs}
\usage{
create_tmh_fasta(proteome_filename, tmhs_filename, tmh_fasta_filename)
}
\arguments{
\item{proteome_filename}{name of the file the proteome will be stored at}

\item{tmhs_filename}{file that contains the TMHs as created by TMHMM}

\item{tmh_fasta_filename}{name of the FASTA file containing only the
TMHs}
}
\value{
nothing, creates file with name \code{tmh_fasta_filename}
}
\description{
Creates a FASTA file with the sequences of the proteome that
are TMHs
}
\author{
Richel J.C. Bilderbeek
}
