% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_tmh_fasta_cpp}
\alias{create_tmh_fasta_cpp}
\title{Creates a FASTA fike with only the TMHs}
\usage{
create_tmh_fasta_cpp(proteome_filename, tmhs_filename, tmh_fasta_filename)
}
\arguments{
\item{proteome_filename}{name of the file the proteome will be stored at}

\item{tmhs_filename}{file that contains the TMHs as created by TMHMM}

\item{tmh_fasta_filename}{name of the FASTA file containing only the TMHs}
}
\value{
nothing
}
\description{
Creates a FASTA fike with only the TMHs
}
\author{
Richel J.C. Bilderbeek
}
