% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{explode}
\alias{explode}
\title{Explodes a string into a character vector}
\usage{
explode(s)
}
\arguments{
\item{s}{one string}
}
\description{
Explodes a string into a character vector
}
\examples{
library(testthat)
x <- explode("ABC")
expect_equal(x, c("A", "B", "C"))
}
\author{
Johannes Textor, Richel J.C. Bilderbeek
}
