% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aa_transition_matrix.R
\name{get_aa_transition_matrix}
\alias{get_aa_transition_matrix}
\title{Get an amino acid transition matrix}
\usage{
get_aa_transition_matrix()
}
\value{
a tibble, with column names equal to the uppercase
  amino acid name
}
\description{
From Dang, Cuong Cao, et al. "FLU, an amino acid substitution model for
influenza proteins." BMC evolutionary biology 10.1 (2010): 99.
}
\details{
File downloaded from \code{ftp://ftp.sanger.ac.uk/pub/1000genomes/lsq/FLU/Flu_All_it2.txt_PAML.txt}

As can be seen in \url{
  https://github.com/stephaneguindon/phyml/blob/master/src/io.c#L3188
}, the column names are \code{Ala Arg Asn Asp Cys Gln Glu Gly His Ile Leu
Lys Met Phe Pro Ser Thr Trp Tyr Val}
}
\author{
Richèl J.C. Bilderbeek
}
