% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adjacent_sequences.R
\name{get_adjacent_sequences}
\alias{get_adjacent_sequences}
\title{Get the adjacent sequences}
\usage{
get_adjacent_sequences(aa_sequence = "VVIILTIAGNILVIMAVSLE")
}
\arguments{
\item{aa_sequence}{amino acid sequence}
}
\description{
Get the adjacent sequences
}
\author{
Richèl J.C. Bilderbeek
}
