% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_amino_acids.R
\name{get_amino_acids}
\alias{get_amino_acids}
\title{Get the twenty standard amino acids, in their one uppercase letter form.}
\usage{
get_amino_acids()
}
\description{
Get the twenty standard amino acids, in their one uppercase letter form.
}
\examples{
library(testthat)

# There are 20 standard amino acids
expect_equal(20, length(get_amino_acids()))

# First one is alanine
expect_equal("A", get_amino_acids()[1])
}
