% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbt_get_paths.R
\name{get_bbbq_paths}
\alias{get_bbbq_paths}
\title{Get the full paths of files in the 'inst/extdata' folder}
\usage{
get_bbbq_paths(filenames)
}
\arguments{
\item{filenames}{the files' names, without the path}
}
\value{
the filenames' full paths, if and only if
  all files are present. Will stop otherwise.
}
\description{
Get the full paths of files in the 'inst/extdata' folder
}
\examples{
  testit::assert(
    length(
      get_bbbq_paths(
        c("UP000001584_83332.fasta", "tmhmm_result.txt")
      )
     ) == 2
   )
}
\seealso{
for one file, use \code{\link{get_bbbq_path}}
}
\author{
Richel J.C. Bilderbeek
}
