% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bergseng_et_al_2015_sup_1.R
\name{get_bergseng_et_al_2015_sup_1}
\alias{get_bergseng_et_al_2015_sup_1}
\title{Get the table of Supplementary Material 1 from
Bergseng et al., 2015.
Will download the file if needed.}
\usage{
get_bergseng_et_al_2015_sup_1(
  url = bbbq::get_bergseng_et_al_2015_sup_1_url(),
  xlsx_filename = file.path(rappdirs::user_data_dir(appname = "bbbq"),
    "251_2014_819_MOESM3_ESM.xlsx"),
  verbose = FALSE
)
}
\arguments{
\item{xlsx_filename}{the XLSX filename}

\item{verbose}{set to TRUE for more output}

\item{URL}{the download URL,
as returned by \link{get_bergseng_et_al_2015_sup_1_url}}
}
\value{
a \link{tibble}[tibble], with columns named
\code{cell_line},
\code{infected},
\code{epitope_sequence},
\code{epitope_length},
\code{copy_number},
\code{source_protein},
\code{gene},
\code{best_allele},
\code{best_allele_rank},
\code{best_allele_predscore},
\code{normalized_copyno}
}
\description{
Get the table of Supplementary Material 1 from
Bergseng et al., 2015.
Will download the file if needed.
}
