% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bergseng_et_al_2015_sup_1_url.R
\name{get_bergseng_et_al_2015_sup_1_url}
\alias{get_bergseng_et_al_2015_sup_1_url}
\title{Get the URL for Bergseng et al., 2015}
\usage{
get_bergseng_et_al_2015_sup_1_url()
}
\description{
Get the URL for Bergseng et al., 2015
}
