% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genetic_code.R
\name{get_genetic_code}
\alias{get_genetic_code}
\title{Get the genetic code}
\usage{
get_genetic_code()
}
\description{
Get the genetic code
}
\examples{
get_genetic_code()
}
\author{
Richèl J.C. Bilderbeek
}
