% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hlas.R
\name{get_hlas}
\alias{get_hlas}
\title{Get the MHC-I HLAs
Copied from 'calculate_overlap'}
\usage{
get_hlas()
}
\description{
Get the MHC-I HLAs
Copied from 'calculate_overlap'
}
\author{
Richèl J.C. Bilderbeek
}
