% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ic50_percentile_binder.R
\name{get_ic50_percentile_binder}
\alias{get_ic50_percentile_binder}
\title{Get the IC50 percentile for which a peptide is a binder}
\usage{
get_ic50_percentile_binder()
}
\description{
Bianchi et al, 2018 used 0.02
}
