% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ic50_threshold.R
\name{get_ic50_threshold}
\alias{get_ic50_threshold}
\title{Get the threshold IC50 value}
\usage{
get_ic50_threshold(peptide_length, haplotype, percentile, ic50_prediction_tool)
}
\arguments{
\item{peptide_length}{length of the peptide in amino acids}

\item{haplotype}{one MHC-I or MHC-II haplotype}

\item{percentile}{how low the IC50 must be for the protein to
be considered a binder. For example, 0.02 denotes that the protein
must have an IC50 in the lowest 2 percent range.
The default value is returned by \link{get_ic50_percentile_binder}.}

\item{ic50_prediction_tool}{tool to predict the IC50 from a
  peptide. Possible values are:\cr
\itemize{
  \item mhcnuggetsr \link[mhcnuggetsr]{mhcnuggetsr},
    which uses MHCnuggets
  \item mhcnuggetsr \link[netmhc2pan]{netmhc2pan},
    which uses NetMHC2pam
  \item EpitopePrediction uses \code{EpitopePrediction}
}}
}
\description{
Get the threshold IC50 value for a certain percentile.
For examples, IC50s below this value may be defined as a binder
}
\author{
Richèl J.C. Bilderbeek
}
