% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ic50s_lut.R
\name{get_ic50s_lut}
\alias{get_ic50s_lut}
\title{Get a lookup table for IC50 predictions of epitopes}
\usage{
get_ic50s_lut(haplotype, peptide_length, ic50_prediction_tool, verbose = TRUE)
}
\description{
Get a lookup table for IC50 predictions of epitopes.
This lookup table is read from file.
If a needed file is absent, it will be downloaded.
Each table is approx 300 megabytes
}
