% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhc1_allele_names.R
\name{get_mhc1_haplotypes}
\alias{get_mhc1_haplotypes}
\title{Deprecated function to the MHC-I allele names used in this study.
Use \link{get_mhc1_allele_names} instead.}
\usage{
get_mhc1_haplotypes()
}
\description{
Deprecated function to the MHC-I allele names used in this study.
Use \link{get_mhc1_allele_names} instead.
}
\author{
Richèl J.C. Bilderbeek
}
