% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhc2_allele_names.R
\name{get_mhc2_allele_names}
\alias{get_mhc2_allele_names}
\title{Get the MHC-II allele names used in this study}
\usage{
get_mhc2_allele_names()
}
\description{
Get the MHC-II allele names used in this study
}
\examples{
get_mhc2_allele_names()
}
\author{
Richèl J.C. Bilderbeek
}
