% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhc2_haplotype_occurance.R
\name{get_mhc2_haplotype_occurance}
\alias{get_mhc2_haplotype_occurance}
\title{Get the MHC-II haplotype occurance}
\usage{
get_mhc2_haplotype_occurance()
}
\value{
a \link{data.frame}
}
\description{
Get the MHC-II haplotype occurance in
both haplotype percentage and phenotype frequency.
}
\details{
Data is from: Greenbaum, Jason, et al. "Functional classification of
class II human leukocyte antigen (HLA) molecules reveals seven different
supertypes and a surprising degree of repertoire sharing across
supertypes." Immunogenetics 63.6 (2011): 325-335.
}
\examples{

df <- get_mhc2_haplotype_occurance()

expect_true("haplotype" \%in\% names(df))
expect_true("haplotype_perc" \%in\% names(df))
expect_true("phenotype_freq" \%in\% names(df))
}
\author{
Richèl J.C. Bilderbeek
}
