% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhc2_allele_names.R
\name{get_mhc2_haplotypes}
\alias{get_mhc2_haplotypes}
\title{Deprecated function to get the MHC-II allele names used in this study.}
\usage{
get_mhc2_haplotypes()
}
\description{
Deprecated function to get the MHC-II allele names used in this study.
Use \link{get_mhc2_allele_names} instead.
}
\examples{
get_mhc2_allele_names()
}
\author{
Richèl J.C. Bilderbeek
}
