% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhc2_ic50_threshold.R
\name{get_mhc2_ic50_threshold}
\alias{get_mhc2_ic50_threshold}
\title{Get the MHC-II IC50 (in nM) threshold}
\usage{
get_mhc2_ic50_threshold()
}
\description{
Get the MHC-II IC50 (in nM) threshold
}
\author{
Richèl J.C. Bilderbeek
}
