% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhc2_peptide_length.R
\name{get_mhc2_peptide_length}
\alias{get_mhc2_peptide_length}
\title{Get the peptide length used in MHC-II epitope predictions}
\usage{
get_mhc2_peptide_length()
}
\description{
Get the peptide length used in MHC-II epitope predictions
}
\examples{
get_mhc2_peptide_length()
}
