% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhc_allele_names.R
\name{get_mhc_haplotypes}
\alias{get_mhc_haplotypes}
\title{Deprecated function to het the MHC allele names used in this study.
Use \link{get_mhc_allele_names} instead}
\usage{
get_mhc_haplotypes()
}
\value{
the MHC allele names as a character vector
}
\description{
Deprecated function to het the MHC allele names used in this study.
Use \link{get_mhc_allele_names} instead
}
\author{
Richèl J.C. Bilderbeek
}
