% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhc_peptide_length.R
\name{get_mhc_peptide_length}
\alias{get_mhc_peptide_length}
\title{Get the MHC peptide length for a specific MHC class}
\usage{
get_mhc_peptide_length(mhc_class)
}
\arguments{
\item{mhc_class}{the MHC class, for example \code{1}}
}
\description{
Get the MHC peptide length for a specific MHC class
}
\examples{
get_mhc_peptide_length(1)
get_mhc_peptide_length(2)
}
