% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_non_tmh_peptides_from_elution.R
\name{get_non_tmh_peptides_from_elution}
\alias{get_non_tmh_peptides_from_elution}
\title{Get the TMH peptides from elution study}
\usage{
get_non_tmh_peptides_from_elution()
}
\description{
Get the TMH peptides from elution study
}
\author{
Richèl J.C. Bilderbeek
}
