% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-protein-lengths.R
\name{get_protein_lengths}
\alias{get_protein_lengths}
\title{Determines the lengths of all proteins in the human proteome.}
\usage{
get_protein_lengths()
}
\description{
Determines the lengths of all proteins in the human proteome.
}
\author{
Richel J.C. Bilderbeek, adapted from Johannes Textor
}
