% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proteome.R
\name{get_proteome}
\alias{get_proteome}
\title{Get the proteome as a list}
\usage{
get_proteome(fasta_filename = system.file("extdata", "short.fasta",
  package = "bbbq"))
}
\arguments{
\item{fasta_filename}{name of a FASTA file}
}
\description{
Get the proteome as a list
}
\author{
Johannes Textor, adapted by Richel J.C. Bilderbeek
}
