% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proteome_filename.R
\name{get_proteome_filename}
\alias{get_proteome_filename}
\title{Get the name of a proteome file}
\usage{
get_proteome_filename(proteome_type = "full", keep_selenoproteins = TRUE)
}
\arguments{
\item{proteome_type}{type of proteome, must be
either \code{full} or \code{representative}}

\item{keep_selenoproteins}{keep the selenoproteins.
The reason to discard these, is that TMHMM cannot deal
with selenoproteins.}
}
\description{
Get the name of a proteome file
}
