% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bbbq_path.R
\name{get_proteome_path}
\alias{get_proteome_path}
\title{Get the path to the proteome}
\usage{
get_proteome_path()
}
\value{
path to the proteome file in \code{inst/extdata}
}
\description{
Get the path to the proteome
}
\author{
Richel J.C. Bilderbeek
}
