% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target_english_name.R
\name{get_target_english_name}
\alias{get_target_english_name}
\title{Get the English description of a target}
\usage{
get_target_english_name(target_name)
}
\arguments{
\item{target_name}{abbreviated name of organism or virus.
Use \link{get_target_names} to get the valid values.
Use \link{get_target_english_name} to get the English name of a target.}
}
\description{
Get the English description of a target
}
