% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_myco_proteome_filename.R
\name{get_test_myco_proteome_filename}
\alias{get_test_myco_proteome_filename}
\title{Get the filename of a test Mycobacterium reference proteome}
\usage{
get_test_myco_proteome_filename()
}
\value{
the filename of the test Mycobacterium reference proteome
}
\description{
Get the filename of a test Mycobacterium reference proteome
}
\author{
Richèl J.C. Bilderbeek
}
