% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_proteome_filename.R
\name{get_test_proteome_filename}
\alias{get_test_proteome_filename}
\title{Get the filename of a testing proteome}
\usage{
get_test_proteome_filename()
}
\value{
the filename of the COVID-19 reference proteome
}
\description{
This testing proteome consists out of on protein with one TMH.
The protein is the viral envelope protein of the COVID-19 reference proteome.
}
\author{
Richèl J.C. Bilderbeek
}
