% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tmh_lengths.R
\name{get_tmh_lengths}
\alias{get_tmh_lengths}
\title{Get the lengths of the TMHs}
\usage{
get_tmh_lengths(epitopeome)
}
\arguments{
\item{epitopeome}{an epitopeome}
}
\value{
a numeric vector of TMH lengths. As a protein can contain
  zero or multiple TMHs, this lengths is probably different than
  the number of proteins
}
\description{
Get the lengths of the TMHs
}
\author{
Richel J.C. Bilderbeek
}
