% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topology.R
\name{get_topology}
\alias{get_topology}
\title{Get a topology as a tibble.}
\usage{
get_topology(
  proteome_type,
  keep_selenoproteins,
  topology_prediction_tool,
  data_folder = rappdirs::user_data_dir(appname = "bbbq")
)
}
\arguments{
\item{proteome_type}{type of proteome, must be
either \code{full} or \code{representative}}

\item{keep_selenoproteins}{keep the selenoproteins.
The reason to discard these, is that TMHMM cannot deal
with selenoproteins.}

\item{topology_prediction_tool}{tool to predict the topology
  of a membrane protein. Possible values are:\cr
\itemize{
  \item mhcnuggetsr \link[pureseqtmr]{pureseqtmr},
    which uses PureseqTM
  \item mhcnuggetsr \link[tmhmm]{tmhmm},
    which uses TMHMM
}}

\item{data_folder}{folder where BBBQ stores its data, such as proteomes
and prediced topologies.}
}
\description{
Topologyies are downloaded and stored at the user's data folder
}
