% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transition_rate.R
\name{get_transition_rate}
\alias{get_transition_rate}
\title{Get the AA transition rate from the FLU transition rate matrix}
\usage{
get_transition_rate(seq_lhs, seq_rhs)
}
\arguments{
\item{seq_lhs}{amino acid to transition from}

\item{seq_rhs}{amino acid to transition to}
}
\description{
Get the AA transition rate from the FLU transition rate matrix
}
