% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transition_rates.R
\name{get_transition_rates}
\alias{get_transition_rates}
\title{Get the transition rates from the one focal protein sequence
to the one or more target protein sequences}
\usage{
get_transition_rates(protein_sequence, protein_sequences)
}
\arguments{
\item{protein_sequence}{a protein sequence}

\item{protein_sequences}{one or more protein sequences}
}
\description{
Get the transition rates from the one focal protein sequence
to the one or more target protein sequences
}
