% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrophobe-controls.R
\name{hydrophobe_controls}
\alias{hydrophobe_controls}
\title{Generates a set of 9mers that do not overlap with transmembrane helices
and have a very similar hydrophobicity distribution as the set of 9mers
that do overlap with transmembrane helices.}
\usage{
hydrophobe_controls()
}
\description{
Generates a set of 9mers that do not overlap with transmembrane helices
and have a very similar hydrophobicity distribution as the set of 9mers
that do overlap with transmembrane helices.
}
\author{
Richel J.C. Bilderbeek, adapted from Johannes Textor
}
