% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrophobicity-distribution.R
\name{hydrophobicity_distribution}
\alias{hydrophobicity_distribution}
\title{Determines the hydrophobicity distribution of all peptides
in the human proteome as well as all peptides overlapping
with transmembrane helices.}
\usage{
hydrophobicity_distribution()
}
\arguments{
\item{tmh_9mers_as_data_filename}{filename to store
per TMH protein, the indices at which it is TMH,
in R data format.
for example `work/tmh.9mers.Rdata`}
}
\description{
Determines the hydrophobicity distribution of all peptides
in the human proteome as well as all peptides overlapping
with transmembrane helices.
}
\author{
Richel J.C. Bilderbeek, adapted from Johannes Textor
}
