% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrophobicity-distribution-elution-data.R
\name{hydrophobicity_distribution_elution_data}
\alias{hydrophobicity_distribution_elution_data}
\title{Something.}
\usage{
hydrophobicity_distribution_elution_data(kyte_doolittle_scale_as_data_filename)
}
\arguments{
\item{kyte_doolittle_scale_as_data_filename}{filename that stores the Kyte-Doolittle scale of
hydrophobicity,
for example `data/kyte.doolittle.scale.Rdata`}
}
\description{
Something.
}
\author{
Richel J.C. Bilderbeek, adapted from Johannes Textor
}
