% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyPref.R
\name{hyPref}
\alias{hyPref}
\title{Does something}
\usage{
hyPref(mhc)
}
\arguments{
\item{mhc}{simplified MHC name, for example, `A02-01`
to use `HLA-A02:01`}
}
\description{
Does something
}
\author{
Richel J.C. Bilderbeek, adapted from Johannes Textor
}
