% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_detected_by_mhc_1.R
\name{is_detected_by_mhc_1}
\alias{is_detected_by_mhc_1}
\title{Determine if MHC-I has at least one binder to the polypeptide}
\usage{
is_detected_by_mhc_1(protein_sequence, mhc_1_haplotype = "HLA-A-02-01")
}
\arguments{
\item{protein_sequence}{a protein sequence}

\item{mhc_1_haplotype}{HMC-I haplotope.
Use \link{get_mhc1_haplotypes} to get a list of all haplotypes}
}
\description{
Determine if MHC-I has at least one binder to the polypeptide
}
