% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_detected_by_mhc_2.R
\name{is_detected_by_mhc_2}
\alias{is_detected_by_mhc_2}
\title{Determine if the  protein sequence has one MHC-II binder}
\usage{
is_detected_by_mhc_2(
  protein_sequence = "VVIIRTIAGRILVIMARSLE",
  mhc_2_haplotype = "HLA-DRB104:05",
  ic50_threshold = get_mhc2_ic50_threshold()
)
}
\arguments{
\item{protein_sequence}{a protein sequence}

\item{mhc_2_haplotype}{HMC-II haplotope.
Use \link{get_mhc2_haplotypes} to get a list of all haplotypes}

\item{ic50_threshold}{peptides with a predicted IC50 value (in nM)
lower than this will be considered binders.
A threshold of 500 (nM) is common.}
}
\description{
Determine if the  protein sequence has one MHC-II binder
}
\author{
Richèl J.C. Bilderbeek
}
