% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_polypeptide.R
\name{is_peptide}
\alias{is_peptide}
\title{Determine if the AA sequence is a peptide}
\usage{
is_peptide(protein_sequence)
}
\arguments{
\item{protein_sequence}{a protein sequence}
}
\description{
Determine if the AA sequence is a peptide
}
\author{
Richèl J.C. Bilderbeek
}
