% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_polypeptide.R
\name{is_polypeptide}
\alias{is_polypeptide}
\title{Determine if the AA sequence is a polypeptide}
\usage{
is_polypeptide(protein_sequence)
}
\arguments{
\item{protein_sequence}{a protein sequence}
}
\description{
Determine if the AA sequence is a polypeptide
}
