% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{nlapply}
\alias{nlapply}
\title{Version of lapply that also passes along the
name of each list element. f must be a function
with arguments n and x.}
\usage{
nlapply(l, f)
}
\arguments{
\item{l}{list}

\item{f}{function}
}
\description{
Version of lapply that also passes along the
name of each list element. f must be a function
with arguments n and x.
}
\author{
Johannes Textor, Richèl J.C. Bilderbeek
}
