% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_markov_chain.R
\name{plot_markov_chain}
\alias{plot_markov_chain}
\title{Show a Markov chain'}
\usage{
plot_markov_chain(
  transition_matrix = matrix(c(0.8, 0.2, 0.1, 0.9), nrow = 2, byrow = TRUE),
  png_filename = tempfile(),
  tool = "tikz"
)
}
\arguments{
\item{transition_matrix}{a transition matrix.
 The first row denotes the transition rates from D.
 All rows must sum up to one.
 \itemize{
   \item \code{[1,1]} denotes \code{D -> D}
   \item \code{[1,2]} denotes \code{D -> U}
   \item \code{[2,1]} denotes \code{U -> D}
   \item \code{[2,2]} denotes \code{U -> U}
}}

\item{png_filename}{name of a PNG file}

\item{tool}{name of the tool used to create the Markov
chain.
Can be \code{tikz}, \code{dot} or \code{igraph}}
}
\description{
Show a Markov chain'
}
