% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_markov_chain.R
\name{plot_markov_chain_dot}
\alias{plot_markov_chain_dot}
\title{Show a Markov chain using DOT}
\usage{
plot_markov_chain_dot(
  transition_matrix = matrix(c(0.5, 0.5, 0.5, 0.5), nrow = 2),
  png_filename = tempfile()
)
}
\arguments{
\item{transition_matrix}{a transition matrix.
 The first row denotes the transition rates from D.
 All rows must sum up to one.
 \itemize{
   \item \code{[1,1]} denotes \code{D -> D}
   \item \code{[1,2]} denotes \code{D -> U}
   \item \code{[2,1]} denotes \code{U -> D}
   \item \code{[2,2]} denotes \code{U -> U}
}}
}
\description{
Show a Markov chain using DOT
}
