% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ic50s.R
\name{predict_ic50s}
\alias{predict_ic50s}
\title{Predict IC50s}
\usage{
predict_ic50s(
  protein_sequence,
  peptide_length,
  haplotype,
  ic50_prediction_tool,
  mhcnuggetsr_peptides_path = mhcnuggetsr::create_temp_peptides_path()
)
}
\arguments{
\item{protein_sequence}{a protein sequence}

\item{peptide_length}{length of the peptide in amino acids}

\item{haplotype}{one MHC-I or MHC-II haplotype}

\item{ic50_prediction_tool}{tool to predict the IC50 from a
  peptide. Possible values are:\cr
\itemize{
  \item mhcnuggetsr \link[mhcnuggetsr]{mhcnuggetsr},
    which uses MHCnuggets
  \item mhcnuggetsr \link[netmhc2pan]{netmhc2pan},
    which uses NetMHC2pam
  \item EpitopePrediction uses \code{EpitopePrediction}
}}

\item{mhcnuggetsr_peptides_path}{path to a temporary file to create
peptides, as needed by \link{mhcnuggetsr},
which is \link[mhcnuggetsr]{create_temp_peptides_path} by default}
}
\description{
Predict IC50s
}
\author{
Richèl J.C. Bilderbeek
}
