% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_n_binders_tmh.R
\name{predict_n_binders_tmh}
\alias{predict_n_binders_tmh}
\title{Calculate the fractions of MHC epitopes that overlap with a TMH
for multiple MHC haplotypes for one type of agent.}
\usage{
predict_n_binders_tmh(target_name, haplotypes, n_aas)
}
\arguments{
\item{haplotypes}{a list of MHC-I or MHC-II haplotypes}

\item{n_aas}{number of amino acids}
}
\value{
a tibble, with columns:
\enumerate{
  \item haplotype haplotype name
  \item n_binders number of epitopes that bind
  \item n_binders_tmh number of epitopes that bind and have one
    amino acid overlapping with a TMH
}
and one row per peptide
}
\description{
Calculate the fractions of MHC epitopes that overlap with a TMH
for multiple MHC haplotypes for one type of agent.
}
\author{
Richèl J.C. Bilderbeek
}
