% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_n_binders_tmh_peptides.R
\name{predict_n_binders_tmh_peptides}
\alias{predict_n_binders_tmh_peptides}
\title{Calculate the fractions of MHC epitopes that overlap with a TMH
for multiple MHC haplotypes for one type of agent.}
\usage{
predict_n_binders_tmh_peptides(
  peptides,
  haplotype,
  peptide_length,
  percentile,
  verbose = FALSE
)
}
\arguments{
\item{peptides}{one ore more peptide sequences}

\item{haplotype}{one MHC-I or MHC-II haplotype}

\item{peptide_length}{length of the peptide in amino acids}

\item{percentile}{how low the IC50 must be for the protein to
be considered a binder. For example, 0.02 denotes that the protein
must have an IC50 in the lowest 2 percent range.
The default value is returned by \link{get_ic50_percentile_binder}.}

\item{verbose}{set to TRUE for more output}
}
\value{
a tibble with columns:
\enumerate{
  \item n_binders number of epitopes that bind
  \item n_binders_tmh number of epitopes that bind and have one
    amino acid overlapping with a TMH
}
and one row per peptide
}
\description{
For one peptide, use \link{predict_n_binders_tmh_peptide}
}
\author{
Richèl J.C. Bilderbeek
}
