% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_n_coincidence_tmh.R
\name{predict_n_coincidence_tmh}
\alias{predict_n_coincidence_tmh}
\title{Calculate the fractions of peptide fragmensts that overlap with a TMH
for one type of agent.}
\usage{
predict_n_coincidence_tmh(target_name, n_aas)
}
\arguments{
\item{target_name}{name of organism or virus or testing
dataset.}

\item{n_aas}{number of amino acids}
}
\value{
a tibble, with columns:
\enumerate{
  \item n_spots number of spots
  \item n_spots_tmh number of spots that have one
    amino acid overlapping with a TMH
}
and one row per peptide
}
\description{
Calculate the fractions of peptide fragmensts that overlap with a TMH
for one type of agent.
}
\author{
Richèl J.C. Bilderbeek
}
