% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_n_coincidence_tmh_peptide.R
\name{predict_n_coincidence_tmh_peptide}
\alias{predict_n_coincidence_tmh_peptide}
\title{Predict the number of n-mers and the number of n-mers
that overlap with at least one amino acid with a TMH
for one peptide.}
\usage{
predict_n_coincidence_tmh_peptide(
  protein_sequence,
  peptide_length,
  verbose = FALSE
)
}
\arguments{
\item{protein_sequence}{a protein sequence}

\item{peptide_length}{length of the peptide in amino acids}

\item{verbose}{set to TRUE for more output}
}
\value{
a tibble with
\enumerate{
  \item n_spots number of spots for the n-mer
  \item n_spots_tmh number of spots that have one
    amino acid overlapping with a TMH
}
}
\description{
Predict the number of n-mers and the number of n-mers
that overlap with at least one amino acid with a TMH
for one peptide.
}
\examples{
library(pureseqtmr)

if (is_pureseqtm_installed()) {

  protein_sequence <- "MYSFVSEETGTLIVNSVLLFLAFVVFLLVTLAILTALRLCAYCCNIVNVSLVKPSFYVYSRV"

  predict_n_coincidence_tmh_peptide(
    protein_sequence = protein_sequence,
    peptide_length = 9
  )
}
}
\author{
Richèl J.C. Bilderbeek
}
