% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_n_coincidence_tmh_peptides.R
\name{predict_n_coincidence_tmh_peptides}
\alias{predict_n_coincidence_tmh_peptides}
\title{Predict the number of n-mers and the number of n-mers
that overlap with at least one amino acid with a TMH
for one peptide.}
\usage{
predict_n_coincidence_tmh_peptides(peptides, peptide_length, verbose = FALSE)
}
\arguments{
\item{peptides}{one ore more peptide sequences}

\item{peptide_length}{length of the peptide in amino acids}

\item{verbose}{set to TRUE for more output}
}
\value{
a tibble with
\enumerate{
  \item n_spots number of spots for the n-mer
  \item n_spots_tmh number of spots that have one
    amino acid overlapping with a TMH
}
and one row per peptide
}
\description{
For one peptide, use \link{predict_n_coincidence_tmh_peptide}
}
\author{
Richèl J.C. Bilderbeek
}
