% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_nth_char.R
\name{replace_nth_char}
\alias{replace_nth_char}
\title{Replace the nth char of a string}
\usage{
replace_nth_char(s, n, c)
}
\arguments{
\item{s}{the string}

\item{n}{the index of the character}

\item{c}{the character to be placed at that index}
}
\description{
Replace the nth char of a string
}
\author{
Richèl J.C. Bilderbeek
}
