% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shred_protein.R
\name{shred_protein}
\alias{shred_protein}
\title{Shred a protein}
\usage{
shred_protein(protein_sequence, peptide_length)
}
\arguments{
\item{protein_sequence}{a protein sequence}

\item{peptide_length}{length of the peptide in amino acids}
}
\description{
Shred a protein
}
