% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_haplotype_names.R
\name{simplify_haplotype_names}
\alias{simplify_haplotype_names}
\title{Simplify haplotype names to numbers}
\usage{
simplify_haplotype_names(haplotype_names)
}
\description{
Simplify haplotype names to numbers
}
