% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_netmhc2pan_name.R
\name{to_netmhc2pan_name}
\alias{to_netmhc2pan_name}
\title{Convert a Greenbaum et al. name to a NetMHC2pan name}
\usage{
to_netmhc2pan_name(name)
}
\arguments{
\item{name}{an allele name as used by Greenbaum et al.}
}
\description{
Convert a Greenbaum et al. name to a NetMHC2pan name
}
