% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_detected.R
\name{are_detected}
\alias{are_detected}
\title{Are the protein sequences detected by either MHC-I or MHC-II?}
\usage{
are_detected(
  protein_sequences,
  mhc_haplotype = "HLA-A*02-01",
  percentile = 0.02
)
}
\arguments{
\item{protein_sequences}{one or more protein sequences}

\item{mhc_haplotype}{haplotype of either HMC-I or MHC-II.
Use \link{get_mhc1_haplotypes} to get a list of all MHC-I haplotypes.
Use \link{get_mhc2_haplotypes} to get a list of all MHC-II haplotypes.}
}
\description{
Are the protein sequences detected by either MHC-I or MHC-II?
}
\author{
Richèl J.C. Bilderbeek
}
