% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_detected.R
\name{are_detected}
\alias{are_detected}
\title{Are the protein sequences detected by either MHC-I or MHC-II?}
\usage{
are_detected(protein_sequences)
}
\arguments{
\item{protein_sequences}{one or more protein sequences}
}
\description{
Are the protein sequences detected by either MHC-I or MHC-II?
}
