% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_detected.R
\name{are_detected}
\alias{are_detected}
\title{Are the protein sequences detected by either MHC-I or MHC-II?}
\usage{
are_detected(protein_sequences, mhc_1_haplotype = "HLA-A-02-01")
}
\arguments{
\item{protein_sequences}{one or more protein sequences}

\item{mhc_1_haplotype}{HMC-I haplotope.
Use \link{get_mhc1_haplotypes} to get a list of all haplotypes}
}
\description{
Are the protein sequences detected by either MHC-I or MHC-II?
}
