% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_p_det_tmh_mut.R
\name{calc_p_det_tmh_mut}
\alias{calc_p_det_tmh_mut}
\title{Calculate the chance that a TMH one-mutation
mutant is detected}
\usage{
calc_p_det_tmh_mut(protein_sequence, mhc_haplotype = "HLA-A-02-01")
}
\arguments{
\item{protein_sequence}{a protein sequence}

\item{mhc_haplotype}{haplotype of either HMC-I or MHC-II.
Use \link{get_mhc1_haplotypes} to get a list of all MHC-I haplotypes.
Use \link{get_mhc2_haplotypes} to get a list of all MHC-II haplotypes.}
}
\description{
Calculate the chance that a TMH one-mutation
mutant is detected
}
\author{
Richèl J.C. Bilderbeek
}
