% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_p_det_tmh_mut.R
\name{calc_p_det_tmh_mut}
\alias{calc_p_det_tmh_mut}
\title{Calculate the chance that a TMH one-mutation
mutant is detected}
\usage{
calc_p_det_tmh_mut(
  protein_sequence,
  mhc_haplotype,
  peptide_length,
  ic50_prediction_tool,
  n_adjancent_sequences = Inf,
  percentile = get_ic50_percentile_binder()
)
}
\arguments{
\item{protein_sequence}{a protein sequence}

\item{mhc_haplotype}{haplotype of either HMC-I or MHC-II.
Use \link{get_mhc1_haplotypes} to get a list of all MHC-I haplotypes.
Use \link{get_mhc2_haplotypes} to get a list of all MHC-II haplotypes.}

\item{peptide_length}{length of the peptide in amino acids}

\item{ic50_prediction_tool}{tool to predict the IC50 from a
  peptide. Possible values are:\cr
\itemize{
  \item mhcnuggetsr \link[mhcnuggetsr]{mhcnuggetsr},
    which uses MHCnuggets
  \item mhcnuggetsr \link[netmhc2pan]{netmhc2pan},
    which uses NetMHC2pam
  \item EpitopePrediction uses \code{EpitopePrediction}
}}

\item{n_adjancent_sequences}{look at the \code{n_adjancent_sequences}
adjacent sequences, instead of all of them. The (maximum) number
of adjacent sequences is \code{19 * peptide_length},
where \code{peptide_length}
is the peptide length in amino acids, and 19 in the number of
possible amino acids to change to (as there are 20 different
amino acids)}

\item{percentile}{how low the IC50 must be for the protein to
be considered a binder. For example, 0.02 denotes that the protein
must have an IC50 in the lowest 2 percent range.
The default value is returned by \link{get_ic50_percentile_binder}.}
}
\description{
Calculate the chance that a TMH one-mutation
mutant is detected
}
\author{
Richèl J.C. Bilderbeek
}
