% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-overlap.R
\name{calculate_overlap}
\alias{calculate_overlap}
\title{Determines the overlap of predicted HLA binders with transmembrane
helices for all HLA supertypes.}
\usage{
calculate_overlap()
}
\arguments{
\item{tmh_9mers_as_data_filename}{filename to store
per TMH protein, the indices at which it is TMH,
in R data format.
for example `work/tmh.9mers.Rdata`}
}
\description{
Determines the overlap of predicted HLA binders with transmembrane
helices for all HLA supertypes.
}
\author{
Richel J.C. Bilderbeek, adapted from Johannes Textor
}
