% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mhc_haplotype_name.R
\name{check_mhc_haplotype_name}
\alias{check_mhc_haplotype_name}
\title{Check an MHC haplotype name}
\usage{
check_mhc_haplotype_name(mhc_haplotype)
}
\description{
Check an MHC haplotype name
}
