% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate-to-hydrophobicity.R
\name{correlate_to_hydrophobicity}
\alias{correlate_to_hydrophobicity}
\title{Correlate to hydrophobicity}
\usage{
correlate_to_hydrophobicity(
  tmh_overlapping_binders_as_data_filename,
  figure_1_d_filename
)
}
\arguments{
\item{tmh_overlapping_binders_as_data_filename}{filename to store
per TMH overlapping binders
in R data format.
for example `work/tmh-overlapping-binders.Rdata`}

\item{figure_1_d_filename}{name of the PDF file to store figure 1d,
for example `plots/figure-1-d.pdf`}
}
\description{
Input: kyte_doolittle_scale_as_data_filename
  and tmh_overlapping_binders_as_data_filename
}
\details{
Output: figure_1_d_filename
}
\author{
Richèl J.C. Bilderbeek, adapted from Johannes Textor
}
