% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_consensus_topology_conservation.R
\name{create_consensus_topology_conservation}
\alias{create_consensus_topology_conservation}
\title{Create a consensus-topology-conservation table}
\usage{
create_consensus_topology_conservation(
  protein_sequences,
  transition_matrix = get_aa_transition_matrix("BLOSUM62")
)
}
\arguments{
\item{protein_sequences}{one or more protein sequences}
}
\value{
a \link{tibble} with columns:\cr
\itemize{
  \item aa the amino acid at that position in consensus sequence
  \item score the conservation score
  \item is_tmh if that amino acid at that location is part of a
    transmembrane helix
}
}
\description{
Create a consensus-topology-conservation table
}
