% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_msa.R
\name{create_msa}
\alias{create_msa}
\title{Creata a multiple-sequence alignment}
\usage{
create_msa(protein_sequences)
}
\arguments{
\item{protein_sequences}{one or more protein sequences}
}
\value{
a \code{MsaAAMultipleAlignment}
}
\description{
Creata a multiple-sequence alignment
}
