% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_extreme_polypeptide.R
\name{create_random_extreme_peptide}
\alias{create_random_extreme_peptide}
\title{Create a random peptide likelier to have an
extreme hydrophobicity}
\usage{
create_random_extreme_peptide(n_aas)
}
\arguments{
\item{n_aas}{number of amino acids}
}
\description{
Create a random peptide likelier to have an
extreme hydrophobicity
}
\author{
Richèl J.C. Bilderbeek
}
