% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_hydrophobic_peptide.R
\name{create_random_hydrophobic_peptide}
\alias{create_random_hydrophobic_peptide}
\title{Create a random hydrophobic peptide}
\usage{
create_random_hydrophobic_peptide(peptide_length)
}
\arguments{
\item{peptide_length}{length of the peptide in amino acids}
}
\description{
Create a random hydrophobic peptide
}
\author{
Richèl J.C. Bilderbeek
}
