% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_polypeptides.R
\name{create_random_peptides}
\alias{create_random_peptides}
\title{Create a random peptide}
\usage{
create_random_peptides(n_peptides, n_aas)
}
\arguments{
\item{n_peptides}{number of peptides}

\item{n_aas}{number of amino acids}
}
\description{
Create a random peptide
}
\author{
Richèl J.C. Bilderbeek
}
