% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_tmh.R
\name{create_random_tmh}
\alias{create_random_tmh}
\title{Create a random TMH sequences}
\usage{
create_random_tmh(
  peptide_length,
  create_random_peptide_fun = bbbq::create_random_hydrophobic_peptide
)
}
\arguments{
\item{peptide_length}{length of the peptide in amino acids}

\item{create_random_peptide_fun}{function to create a random
peptide fragment of the desired length}
}
\description{
Create a random TMH sequences
}
\author{
Richèl J.C. Bilderbeek
}
