% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_tmh_fasta_cpp}
\alias{create_tmh_fasta_cpp}
\title{Calculate the auto correlation time
from \url{https://github.com/beast-dev/beast-mcmc/blob/800817772033c13061f026226e41128d21fd14f3/src/dr/inference/trace/TraceCorrelation.java#L159} # nolint}
\usage{
create_tmh_fasta_cpp(proteome_filename, tmhs_filename, tmh_fasta_filename)
}
\arguments{
\item{sample}{sample}

\item{sample_interval}{sample interval}
}
\value{
the auto correlation time
}
\description{
Calculate the auto correlation time
from \url{https://github.com/beast-dev/beast-mcmc/blob/800817772033c13061f026226e41128d21fd14f3/src/dr/inference/trace/TraceCorrelation.java#L159} # nolint
}
\author{
Richel J.C. Bilderbeek
}
