% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(fasta_filename, proteome_filename, tmh_fasta_filename,
  tmhs_filename, verbose)
}
\arguments{
\item{fasta_filename}{name of a FASTA file}

\item{proteome_filename}{name of the file the proteome will be stored at}

\item{tmh_fasta_filename}{name of the FASTA file containing only the
TMHs}

\item{tmhs_filename}{file that contains the TMHs as created by TMHMM}

\item{verbose}{set to TRUE for more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richel J.C. Bilderbeek
}
