% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  binding_strength_threshold,
  epitopeome_filename,
  fasta_filename,
  haplotype,
  haplotypes,
  hydrophobe_control_peptides_as_data_filename,
  ic50_prediction_tool,
  ic50_threshold,
  kyte_doolittle_scale_as_data_filename,
  max_hydrophobicity,
  mhc_class,
  mhc_haplotype,
  mhc_1_haplotype,
  mhc_2_haplotype,
  min_hydrophobicity,
  msa_method,
  msa_subst_matrix,
  n_peptides,
  non_tmh_eluted_filename,
  peptide,
  peptides,
  peptide_length,
  percentile,
  png_filename,
  protein_lengths_filename,
  protein_sequence,
  protein_sequences,
  proteome_filename,
  proteome_as_data_filename,
  proteome_9mer_hydrophobicity_as_data_filename,
  target_name,
  tmh_9mers_as_data_filename,
  tmh_eluted_filename,
  tmh_overlapping_binders_as_data_filename,
  pureseqtmr_filename,
  topology,
  topology_prediction_tool,
  transition_matrix,
  transition_matrix_name,
  trans_membrane_analysis_filename,
  triplet,
  verbose
)
}
\arguments{
\item{binding_strength_threshold}{threshold of the MHC2 allele's
binding strength. A low threshold, e.g. \code{5.0}, selects for the top
5 percent best binders}

\item{epitopeome_filename}{name of the file to store the epitopeome
output.
By default, this filename is generated by \link{tempfile},
and will be put in the \code{/tmp} folder,
which is cleaned by the operating system}

\item{fasta_filename}{name of a FASTA file}

\item{haplotype}{one MHC-I or MHC-II haplotype}

\item{haplotypes}{a list of MHC-I or MHC-II haplotypes}

\item{hydrophobe_control_peptides_as_data_filename}{filename that stores hydrophobe control peptides
as R data,
for example `work/hydrophobe-control-peptides.Rdata`}

\item{ic50_prediction_tool}{tool to predict the IC50 from a
  peptide. Possible values are:\cr
\itemize{
  \item mhcnuggetsr \link[mhcnuggetsr]{mhcnuggetsr},
    which uses MHCnuggets
  \item mhcnuggetsr \link[netmhc2pan]{netmhc2pan},
    which uses NetMHC2pam
  \item EpitopePrediction uses \code{EpitopePrediction}
}}

\item{ic50_threshold}{peptides with a predicted IC50 value (in nM)
lower than this will be considered binders.
A threshold of 500 (nM) is common.}

\item{kyte_doolittle_scale_as_data_filename}{filename that stores the Kyte-Doolittle scale of
hydrophobicity,
for example `data/kyte.doolittle.scale.Rdata`}

\item{max_hydrophobicity}{the maximal hydrophobicity}

\item{mhc_class}{the MHC class, for example \code{1}}

\item{mhc_haplotype}{haplotype of either HMC-I or MHC-II.
Use \link{get_mhc1_haplotypes} to get a list of all MHC-I haplotypes.
Use \link{get_mhc2_haplotypes} to get a list of all MHC-II haplotypes.}

\item{mhc_1_haplotype}{HMC-I haplotope.
Use \link{get_mhc1_haplotypes} to get a list of all haplotypes}

\item{mhc_2_haplotype}{HMC-II haplotope.
Use \link{get_mhc2_haplotypes} to get a list of all haplotypes}

\item{min_hydrophobicity}{the minimal hydrophobicity}

\item{msa_method}{the \code{method} argument of \link[msa]{msa}.
Valid methods are \code{ClustalW}, \code{ClustalOmega} (preferred for proteins)
and \code{Muscle}}

\item{msa_subst_matrix}{multiple sequence alignment substitution matrix,
the \code{substitutionMatrix} argument of \link[msa]{msa}.
For a protein alignment, relevant values are \code{blosum},
\code{pam} and \code{gonnet}}

\item{n_peptides}{number of peptides}

\item{non_tmh_eluted_filename}{Filename for the non-TMHs found on B-cells,
for example `data/non-TMH-Bcell-elution.txt`}

\item{peptide}{peptide sequence, for example,
\code{ARNDCQEV}}

\item{peptides}{one ore more peptide sequences}

\item{peptide_length}{length of the peptide in amino acids}

\item{percentile}{how low the IC50 must be for the protein to
be considered a binder. For example, 0.02 denotes that the protein
must have an IC50 in the lowest 2 percent range.
The default value is returned by \link{get_ic50_percentile_binder}.}

\item{png_filename}{name of a PNG file}

\item{protein_lengths_filename}{filename to store the
length of all proteins in proteome,
for example `work/protein-lengths.txt`}

\item{protein_sequence}{a protein sequence}

\item{protein_sequences}{one or more protein sequences}

\item{proteome_filename}{name of the file the proteome will be stored at}

\item{proteome_as_data_filename}{filename to store the
proteome in R data format,
for example `work/proteome.Rdata`}

\item{proteome_9mer_hydrophobicity_as_data_filename}{filename to store the
proteome 9-mer hydrophobicities in R data format,
for example `work/proteome.9mer.hydrophobicity.Rdata`}

\item{target_name}{name of organism or virus or testing
dataset.}

\item{tmh_9mers_as_data_filename}{filename to store
per TMH protein, the indices at which it is TMH,
in R data format.
for example `work/tmh.9mers.Rdata`}

\item{tmh_eluted_filename}{Filename for the TMHs found on B-cells,
for example, `data/TMH-Bcell-elution.txt`}

\item{tmh_overlapping_binders_as_data_filename}{filename to store
per TMH overlapping binders
in R data format.
for example `work/tmh-overlapping-binders.Rdata`}

\item{pureseqtmr_filename}{name of the file to store the
  PureseqTM output.
  By default, this filename is generated by \link{tempfile},
  and will be put in the \code{/tmp} folder,
  which is cleaned by the operating system
\code{mhcn} for \link[mhcnuggetsr]{mhcnuggetsr}}

\item{topology}{a topology,
as created by \link[pureseqtmr]{predict_topology}}

\item{topology_prediction_tool}{tool to predict the topology
  of a membrane protein. Possible values are:\cr
\itemize{
  \item mhcnuggetsr \link[pureseqtmr]{pureseqtmr},
    which uses PureseqTM
  \item mhcnuggetsr \link[tmhmm]{tmhmm},
    which uses TMHMM
}}

\item{transition_matrix}{the amino acid transition matrix,
for example, the BLOSUM62 transition matrix
as returned by \link{get_aa_transition_matrix}}

\item{transition_matrix_name}{the name of an amino acid transition matrix,
for example, \code{BLOSUM62}}

\item{trans_membrane_analysis_filename}{Filename for
uhhh, something, for example
`tmh-predictions/trans-membrane-analysis-shortened.txt`}

\item{triplet}{a DNA triplet}

\item{verbose}{set to TRUE for more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
