% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  alleles,
  binding_strength_threshold,
  epitopeome_filename,
  fasta_filename,
  haplotypes,
  hydrophobe_control_peptides_as_data_filename,
  kyte_doolittle_scale_as_data_filename,
  n_aas,
  netmhc2pan_filename,
  non_tmh_eluted_filename,
  png_filename,
  protein_lengths_filename,
  protein_sequence,
  protein_sequences,
  proteome_filename,
  proteome_as_data_filename,
  proteome_9mer_hydrophobicity_as_data_filename,
  tmh_9mers_as_data_filename,
  tmh_eluted_filename,
  tmh_overlapping_binders_as_data_filename,
  tmhmm_filename,
  trans_membrane_analysis_filename,
  triplet,
  verbose
)
}
\arguments{
\item{alleles}{one or more alleles, e.g. \code{DRB1_0101}.
See \link[netmhc2pan]{get_netmhc2pan_alleles} for a full list}

\item{binding_strength_threshold}{threshold of the MHC2 allele's
binding strength. A low threshold, e.g. \code{5.0}, selects for the top
5 percent best binders}

\item{epitopeome_filename}{name of the file to store the epitopeome
output.
By default, this filename is generated by \link{tempfile},
and will be put in the \code{/tmp} folder,
which is cleaned by the operating system}

\item{fasta_filename}{name of a FASTA file}

\item{haplotypes}{a list of MHC-I or MHC-II haplotypes}

\item{hydrophobe_control_peptides_as_data_filename}{filename that stores hydrophobe control peptides
as R data,
for example `work/hydrophobe-control-peptides.Rdata`}

\item{kyte_doolittle_scale_as_data_filename}{filename that stores the Kyte-Doolittle scale of
hydrophobicity,
for example `data/kyte.doolittle.scale.Rdata`}

\item{n_aas}{number of amino acids}

\item{netmhc2pan_filename}{name of the file to store the
NetMHC2pan output.
By default, this filename is generated by \link{tempfile},
and will be put in the \code{/tmp} folder,
which is cleaned by the operating system}

\item{non_tmh_eluted_filename}{Filename for the non-TMHs found on B-cells,
for example `data/non-TMH-Bcell-elution.txt`}

\item{png_filename}{name of a PNG file}

\item{protein_lengths_filename}{filename to store the
length of all proteins in proteome,
for example `work/protein-lengths.txt`}

\item{protein_sequence}{a protein sequence}

\item{protein_sequences}{one or more protein sequences}

\item{proteome_filename}{name of the file the proteome will be stored at}

\item{proteome_as_data_filename}{filename to store the
proteome in R data format,
for example `work/proteome.Rdata`}

\item{proteome_9mer_hydrophobicity_as_data_filename}{filename to store the
proteome 9-mer hydrophobicities in R data format,
for example `work/proteome.9mer.hydrophobicity.Rdata`}

\item{tmh_9mers_as_data_filename}{filename to store
per TMH protein, the indices at which it is TMH,
in R data format.
for example `work/tmh.9mers.Rdata`}

\item{tmh_eluted_filename}{Filename for the TMHs found on B-cells,
for example, `data/TMH-Bcell-elution.txt`}

\item{tmh_overlapping_binders_as_data_filename}{filename to store
per TMH overlapping binders
in R data format.
for example `work/tmh-overlapping-binders.Rdata`}

\item{tmhmm_filename}{name of the file to store the
TMHMM output.
By default, this filename is generated by \link{tempfile},
and will be put in the \code{/tmp} folder,
which is cleaned by the operating system}

\item{trans_membrane_analysis_filename}{Filename for
uhhh, something, for example
`tmh-predictions/trans-membrane-analysis-shortened.txt`}

\item{triplet}{a DNA triplet}

\item{verbose}{set to TRUE for more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richel J.C. Bilderbeek
}
