% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{explode}
\alias{explode}
\title{Explodes a string into a character vector}
\usage{
explode(s)
}
\arguments{
\item{s}{one string}
}
\description{
Explodes a string into a character vector
}
\examples{
# c("A", "B", "C")
explode("ABC")
}
\author{
Johannes Textor, Richèl J.C. Bilderbeek
}
