% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aa_transition_matrix.R
\name{get_aa_transition_matrix}
\alias{get_aa_transition_matrix}
\title{Get an amino acid transition matrix}
\usage{
get_aa_transition_matrix()
}
\description{
From Dang, Cuong Cao, et al. "FLU, an amino acid substitution model for
influenza proteins." BMC evolutionary biology 10.1 (2010): 99.
}
\details{
File downloaded from \code{ftp://ftp.sanger.ac.uk/pub/1000genomes/lsq/FLU/Flu_All_it2.txt_PAML.txt}
}
