% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhc2_haplotypes.R
\name{get_mhc2_haplotypes}
\alias{get_mhc2_haplotypes}
\title{Get the MHC-II haplotypes used in this study}
\usage{
get_mhc2_haplotypes()
}
\description{
Get the MHC-II haplotypes used in this study
}
\author{
Richèl J.C. Bilderbeek
}
