% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_lengths.R
\name{get_protein_lengths}
\alias{get_protein_lengths}
\title{Get an overview of all proteins and their lengths in number of
amino acids.}
\usage{
get_protein_lengths(fasta_filename = system.file("extdata",
  "short.fasta", package = "bbbq"))
}
\arguments{
\item{fasta_filename}{name of a FASTA file}
}
\value{
a data frame with each protein's name and the number'
  of amino acids it consists of
}
\description{
Get an overview of all proteins and their lengths in number of
amino acids.
}
\author{
Richel J.C. Bilderbeek
}
