% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proteome.R
\name{get_proteome}
\alias{get_proteome}
\title{Get a proteome as a tibble.}
\usage{
get_proteome(
  proteome_type = "full",
  keep_selenoproteins = TRUE,
  data_folder = rappdirs::user_data_dir(appname = "bbbq")
)
}
\arguments{
\item{proteome_type}{type of proteome, must be
either \code{full} or \code{representative}}

\item{keep_selenoproteins}{keep the selenoproteins.
The reason to discard these, is that TMHMM cannot deal
with selenoproteins.}

\item{data_folder}{folder where BBBQ stores its data, such as proteomes
and prediced topologies.}
}
\description{
Proteomes are downloaded and stored at the user's data folder
}
