% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topology.R
\name{get_topology}
\alias{get_topology}
\title{Get a topology as a tibble.}
\usage{
get_topology(
  target_name = "human",
  proteome_type = "full",
  keep_selenoproteins = TRUE,
  topology_prediction_tool,
  data_folder = rappdirs::user_data_dir(appname = "bbbq"),
  verbose = FALSE
)
}
\arguments{
\item{target_name}{abbreviated name of organism or virus.
Use \link{get_target_names} to get the valid values.
Use \link{get_target_english_name} to get the English name of a target.}

\item{proteome_type}{type of proteome, must be
either \code{full} or \code{representative}}

\item{keep_selenoproteins}{keep the selenoproteins.
The reason to discard these, is that TMHMM cannot deal
with selenoproteins.}

\item{topology_prediction_tool}{tool to predict the topology
  of a membrane protein. Possible values are:\cr
\itemize{
  \item mhcnuggetsr \link[pureseqtmr]{pureseqtmr},
    which uses PureseqTM
  \item mhcnuggetsr \link[tmhmm]{tmhmm},
    which uses TMHMM
}}

\item{data_folder}{folder where BBBQ stores its data, such as proteomes
and prediced topologies.}

\item{verbose}{set to TRUE for more output}
}
\description{
Topologyies are downloaded and stored at the user's data folder
}
