% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transition_rate.R
\name{get_transition_rate}
\alias{get_transition_rate}
\title{Get the AA transition rate from the FLU transition rate matrix}
\usage{
get_transition_rate(a, b)
}
\arguments{
\item{a}{amino acid to transition from}

\item{b}{amino acid to transition to}
}
\description{
Get the AA transition rate from the FLU transition rate matrix
}
