% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uniprot_id.R
\name{get_uniprot_id}
\alias{get_uniprot_id}
\title{Get the Uniprot ID from a target}
\usage{
get_uniprot_id(target_name)
}
\arguments{
\item{target_name}{name of organism or virus or testing
dataset.}
}
\description{
Get the Uniprot ID from a target
}
