% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-overlap-controls.R
\name{hydrophobe_overlap_controls}
\alias{hydrophobe_overlap_controls}
\title{Performs the same analysis as "calculate-overlap.R", but now for the
'control population' consisting of 9mers that do not overlap with
transmembrane helices.}
\usage{
hydrophobe_overlap_controls()
}
\description{
Performs the same analysis as "calculate-overlap.R", but now for the
'control population' consisting of 9mers that do not overlap with
transmembrane helices.
}
\author{
Richel J.C. Bilderbeek, adapted from Johannes Textor
}
