% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrophobicity-distribution.R
\name{hydrophobicity_distribution}
\alias{hydrophobicity_distribution}
\title{Determines the hydrophobicity distribution of all peptides
in the human proteome as well as all peptides overlapping
with transmembrane helices.}
\usage{
hydrophobicity_distribution(
  proteome_as_data_filename,
  tmh_9mers_as_data_filename,
  kyte_doolittle_scale_as_data_filename,
  proteome_9mer_hydrophobicity_as_data_filename
)
}
\arguments{
\item{proteome_as_data_filename}{filename to store the
proteome in R data format,
for example `work/proteome.Rdata`}

\item{tmh_9mers_as_data_filename}{filename to store
per TMH protein, the indices at which it is TMH,
in R data format.
for example `work/tmh.9mers.Rdata`}

\item{kyte_doolittle_scale_as_data_filename}{filename that stores the Kyte-Doolittle scale of
hydrophobicity,
for example `data/kyte.doolittle.scale.Rdata`}

\item{proteome_9mer_hydrophobicity_as_data_filename}{filename to store the
proteome 9-mer hydrophobicities in R data format,
for example `work/proteome.9mer.hydrophobicity.Rdata`}
}
\description{
Inputs:
  * proteome_as_data_filename
  * tmh_9mers_as_data_filename
  * kyte_doolittle_scale_as_data_filename
}
\details{
Output:
  * proteome_9mer_hydrophobicity_as_data_filename
}
\author{
Richel J.C. Bilderbeek, adapted from Johannes Textor
}
