% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrophobicity-distribution-elution-data.R
\name{hydrophobicity_distribution_elution_data}
\alias{hydrophobicity_distribution_elution_data}
\title{Creates figures 4a and 4b}
\usage{
hydrophobicity_distribution_elution_data(
  figure_3_a_filename,
  figure_3_b_filename
)
}
\arguments{
\item{figure_3_a_filename}{name of the PDF for figure 4a,
for example `plots/figure-3-a.pdf`}

\item{figure_3_b_filename}{name of the PDF for figure 4b,
for example `plots/figure-3-b.pdf`}
}
\description{
Note that in the original version, these were called
3a and 3b in the code.
}
\author{
Richèl J.C. Bilderbeek, adapted from Johannes Textor
}
