% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrophobicity-distribution-elution-data.R
\name{hydrophobicity_distribution_elution_data}
\alias{hydrophobicity_distribution_elution_data}
\title{Something.}
\usage{
hydrophobicity_distribution_elution_data(
  kyte_doolittle_scale_as_data_filename,
  tmh_eluted_filename,
  non_tmh_eluted_filename
)
}
\arguments{
\item{kyte_doolittle_scale_as_data_filename}{filename that stores the Kyte-Doolittle scale of
hydrophobicity,
for example `data/kyte.doolittle.scale.Rdata`}

\item{tmh_eluted_filename}{Filename for the TMHs found on B-cells,
for example, `data/TMH-Bcell-elution.txt`}

\item{non_tmh_eluted_filename}{Filename for the non-TMHs found on B-cells,
for example `data/non-TMH-Bcell-elution.txt`}
}
\description{
Inputs:
 * kyte_doolittle_scale_as_data_filename
 * tmh_eluted_filename
 * non_tmh_eluted_filename
}
\author{
Richel J.C. Bilderbeek, adapted from Johannes Textor
}
