% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_detected.R
\name{is_detected}
\alias{is_detected}
\title{Determine if MHC-I or MHC-II has at least one binder to the polypeptide}
\usage{
is_detected(protein_sequence)
}
\arguments{
\item{protein_sequence}{a protein sequence}
}
\description{
Determine if MHC-I or MHC-II has at least one binder to the polypeptide
}
