% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_detected.R
\name{is_detected}
\alias{is_detected}
\title{Determine if the haplotype has at least one binder to the peptide}
\usage{
is_detected(protein_sequence, mhc_haplotype, percentile = 0.02)
}
\arguments{
\item{protein_sequence}{a protein sequence}

\item{mhc_haplotype}{haplotype of either HMC-I or MHC-II.
Use \link{get_mhc1_haplotypes} to get a list of all MHC-I haplotypes.
Use \link{get_mhc2_haplotypes} to get a list of all MHC-II haplotypes.}

\item{percentile}{how low the IC50 must be for the protein to
be considered a binder. For example, 0.02 denotes that the protein
must have an IC50 in the lowest 2 percent range}
}
\description{
Determine if the haplotype has at least one binder to the peptide
}
\author{
Richèl J.C. Bilderbeek
}
