% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_detected_by_mhc_2.R
\name{is_detected_by_mhc_2}
\alias{is_detected_by_mhc_2}
\title{Determine if the  protein sequence has one MHC-II binder}
\usage{
is_detected_by_mhc_2(protein_sequence = "VVIIRTIAGRILVIMARSLE")
}
\arguments{
\item{protein_sequence}{a protein sequence}
}
\description{
Determine if the  protein sequence has one MHC-II binder
}
