% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ic50s_mhc1.R
\name{predict_ic50s_mhc1}
\alias{predict_ic50s_mhc1}
\title{Predict the IC50s from one peptide}
\usage{
predict_ic50s_mhc1(
  protein_sequence = "VVIILTIAGNILVIMAVSLE",
  mhc_1_haplotype,
  n_aas,
  tool = "ep"
)
}
\arguments{
\item{protein_sequence}{a protein sequence}

\item{mhc_1_haplotype}{HMC-I haplotope.
Use \link{get_mhc1_haplotypes} to get a list of all haplotypes}

\item{n_aas}{number of amino acids}

\item{tool}{the software tool used.
Can be \code{ep} for \code{EpitopePrediction} or
\code{mhcn} for \link[mhcnuggetsr]{mhcnuggetsr}}
}
\description{
Predict the IC50s from one peptide
}
\examples{
library(testthat)

n_aas <- 9
df <- predict_ic50s_mhc1(
  protein_sequence = "VVIILTIAGNILVIMAVSLE",
  mhc_1_haplotype = get_mhc1_haplotypes()[1],
  n_aas = n_aas,
  tool = "mhcn"
)
expect_true("peptide" \%in\% names(df))
expect_true("ic50" \%in\% names(df))
expect_true(all(nchar(df$peptide) == n_aas))
}
