% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_n_binders_tmh_peptide.R
\name{predict_n_binders_tmh_peptide}
\alias{predict_n_binders_tmh_peptide}
\title{Predict the number of binders and the number of binders
that overlap with at least one amino acid with a TMH
for one peptide.}
\usage{
predict_n_binders_tmh_peptide(
  peptide,
  haplotype,
  n_aas,
  percentile,
  verbose = FALSE
)
}
\arguments{
\item{peptide}{peptide sequence, for example,
\code{ARNDCQEV}}

\item{haplotype}{one MHC-I or MHC-II haplotype}

\item{n_aas}{number of amino acids}

\item{percentile}{how low the IC50 must be for the protein to
be considered a binder. For example, 0.02 denotes that the protein
must have an IC50 in the lowest 2 percent range}

\item{verbose}{set to TRUE for more output}
}
\value{
a tibble with
\enumerate{
  \item n_binders number of epitopes that bind
  \item n_binders_tmh number of epitopes that bind and have one
    amino acid overlapping with a TMH
}
}
\description{
Predict the number of binders and the number of binders
that overlap with at least one amino acid with a TMH
for one peptide.
}
\examples{
library(mhcnuggetsr)
library(pureseqtmr)

if (is_mhcnuggets_installed() && is_pureseqtm_installed()) {

  peptide <- "MYSFVSEETGTLIVNSVLLFLAFVVFLLVTLAILTALRLCAYCCNIVNVSLVKPSFYVYSRV"

  predict_n_binders_tmh_peptide(
    peptide = peptide,
    haplotype = get_mhc1_haplotypes()[1],
    n_aas = 9,
  )
}
}
\author{
Richèl J.C. Bilderbeek
}
