% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_n_binders_tmh_peptide.R
\name{predict_n_binders_tmh_peptide}
\alias{predict_n_binders_tmh_peptide}
\title{Predict the number of binders and the number of binders
that overlap with at least one amino acid with a TMH
for one peptide.}
\usage{
predict_n_binders_tmh_peptide(
  protein_sequence,
  haplotype,
  peptide_length,
  percentile,
  verbose = FALSE,
  ic50_prediction_tool = "mhcnuggetsr"
)
}
\arguments{
\item{protein_sequence}{a protein sequence}

\item{haplotype}{one MHC-I or MHC-II haplotype}

\item{peptide_length}{length of the peptide in amino acids}

\item{percentile}{how low the IC50 must be for the protein to
be considered a binder. For example, 0.02 denotes that the protein
must have an IC50 in the lowest 2 percent range.
The default value is returned by \link{get_ic50_percentile_binder}.}

\item{verbose}{set to TRUE for more output}

\item{ic50_prediction_tool}{tool to predict the IC50 from a
  peptide. Possible values are:\cr
\itemize{
  \item mhcnuggetsr \link[mhcnuggetsr]{mhcnuggetsr},
    which uses MHCnuggets
  \item mhcnuggetsr \link[netmhc2pan]{netmhc2pan},
    which uses NetMHC2pam
  \item EpitopePrediction uses \code{EpitopePrediction}
}}
}
\value{
a tibble with
\enumerate{
  \item n_binders number of epitopes that bind
  \item n_binders_tmh number of epitopes that bind and have one
    amino acid overlapping with a TMH
}
}
\description{
Predict the number of binders and the number of binders
that overlap with at least one amino acid with a TMH
for one peptide.
}
\examples{
library(mhcnuggetsr)
library(pureseqtmr)

if (is_mhcnuggets_installed() && is_pureseqtm_installed()) {

  protein_sequence <- "MYSFVSEETGTLIVNSVLLFLAFV"

  predict_n_binders_tmh_peptide(
    protein_sequence = protein_sequence,
    haplotype = get_mhc1_haplotypes()[1],
    peptide_length = 9,
    percentile = 0.123
  )
}
}
\author{
Richèl J.C. Bilderbeek
}
