% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_n_binders_tmh_peptides.R
\name{predict_n_binders_tmh_peptides}
\alias{predict_n_binders_tmh_peptides}
\title{Calculate the fractions of MHC epitopes that overlap with a TMH
for multiple MHC haplotypes for one type of agent.}
\usage{
predict_n_binders_tmh_peptides(
  peptides,
  target_name,
  haplotype,
  n_aas,
  verbose = FALSE
)
}
\arguments{
\item{n_aas}{number of amino acids}

\item{verbose}{set to TRUE for more output}
}
\value{
a tibble with columns:
\enumerate{
  \item n_binders number of epitopes that bind
  \item n_binders_tmh number of epitopes that bind and have one
    amino acid overlapping with a TMH
}
and one row per peptide
}
\description{
For one peptide, use \link{predict_n_binders_tmh_peptide}
}
\author{
Richèl J.C. Bilderbeek
}
