% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_n_coincidence_tmh_peptide.R
\name{predict_n_coincidence_tmh_peptide}
\alias{predict_n_coincidence_tmh_peptide}
\title{Predict the number of n-mers and the number of n-mers
that overlap with at least one amino acid with a TMH
for one peptide.}
\usage{
predict_n_coincidence_tmh_peptide(peptide, n_aas, verbose = FALSE)
}
\arguments{
\item{peptide}{peptide sequence, for example,
\code{ARNDCQEV}}

\item{n_aas}{number of amino acids}

\item{verbose}{set to TRUE for more output}
}
\value{
a tibble with
\enumerate{
  \item n_spots number of spots for the n-mer
  \item n_spots_tmh number of spots that have one
    amino acid overlapping with a TMH
}
}
\description{
Predict the number of n-mers and the number of n-mers
that overlap with at least one amino acid with a TMH
for one peptide.
}
\examples{
library(pureseqtmr)

if (is_pureseqtm_installed()) {

  peptide <- "MYSFVSEETGTLIVNSVLLFLAFVVFLLVTLAILTALRLCAYCCNIVNVSLVKPSFYVYSRV"

  predict_n_coincidence_tmh_peptide(
    peptide = peptide,
    n_aas = 9
  )
}
}
\author{
Richèl J.C. Bilderbeek
}
