% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Converts the human proteome, as downloaded from UniProt,
into a more convenient data structure to work with: a named
list of strings.}
\usage{
prepare_data(fasta_filename = system.file("extdata", "short.fasta",
  package = "bbbq"), proteome_filename = tempfile("proteome.Rdata"))
}
\arguments{
\item{fasta_filename}{name of a FASTA file}

\item{proteome_filename}{name of the file the proteome will be stored at}
}
\description{
Converts the human proteome, as downloaded from UniProt,
into a more convenient data structure to work with: a named
list of strings.
}
\author{
Johannes Textor, adapted by Richel J.C. Bilderbeek
}
