% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{prepare_data}
\alias{prepare_data}
\title{Converts the human proteome, as downloaded from UniProt,
into a more convenient data structure to work with: a named
list of strings.}
\usage{
prepare_data(
  fasta_filename,
  trans_membrane_analysis_filename,
  protein_lengths_filename,
  proteome_as_data_filename,
  tmh_9mers_as_data_filename
)
}
\arguments{
\item{fasta_filename}{proteome as FASTA file,
for example `proteome/UP000005640_9606.fasta.gz`:}

\item{trans_membrane_analysis_filename}{Filename for
uhhh, something, for example
`tmh-predictions/trans-membrane-analysis-shortened.txt`}

\item{protein_lengths_filename}{filename to store the
length of all proteins in proteome,
for example `work/protein-lengths.txt`}

\item{proteome_as_data_filename}{filename to store the
proteome in R data format,
for example `work/proteome.Rdata`}

\item{tmh_9mers_as_data_filename}{filename to store
per TMH protein, the indices at which it is TMH,
in R data format.
for example `work/tmh.9mers.Rdata`}
}
\description{
Also creates a few other files that are helpful for the analysis.
}
\details{
Proteome is downloaded from:

  ftp://ftp.ebi.ac.uk/pub/databases/reference_proteomes/
    QfO/Eukaryota/UP000005640_9606.fasta.gz

Proteome used in Bianchi et al., 2017 used:

  https://github.com/richelbilderbeek/bianchi_et_al_2017/raw/master/proteome_2017/UP000005640_9606.fasta.gz
}
\author{
Richèl J.C. Bilderbeek, adapted from Johannes Textor
}
