% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{prepare_data}
\alias{prepare_data}
\title{Converts the human proteome, as downloaded from UniProt,
into a more convenient data structure to work with: a named
list of strings.}
\usage{
prepare_data(fasta_filename)
}
\arguments{
\item{fasta_filename}{proteome as FASTA file,
for example `proteome/UP000005640_9606.fasta.gz`:}
}
\description{
Also creates a few other files that are helpful for the analysis.
}
\details{
Output files:

 * `work/protein-lengths.txt`: length of all proteins in proteome
 * `work/proteome.Rdata`: proteome as an R list
 * `work/tmh.9mers.Rdata`: per TMH protein, the indices at which it is TMH

Proteome is downloaded from:

  ftp://ftp.ebi.ac.uk/pub/databases/reference_proteomes/
    QfO/Eukaryota/UP000005640_9606.fasta.gz
}
\author{
Richel J.C. Bilderbeek, adapted from Johannes Textor
}
