% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_epitope_prediction_name.R
\name{to_epitope_prediction_name}
\alias{to_epitope_prediction_name}
\title{Convert an academic HMC-I haplotype name to one to be used by
the EpitopePrediction package}
\usage{
to_epitope_prediction_name(mhc_1_haplotype)
}
\arguments{
\item{mhc_1_haplotype}{HMC-I haplotope.
Use \link{get_mhc1_haplotypes} to get a list of all haplotypes}
}
\description{
Convert an academic HMC-I haplotype name to one to be used by
the EpitopePrediction package
}
